Refactor monolithical service.rs into smaller mods in lib
